forgot to actually set the uk6
