+++ library/src/main/java/com/pokegoapi/main/AsyncServerRequest.java
@@ -18,6 +18,7 @@ package com.pokegoapi.main;
 import POGOProtos.Networking.Requests.RequestOuterClass;
 import POGOProtos.Networking.Requests.RequestTypeOuterClass;
 import com.google.protobuf.GeneratedMessage;
+import com.pokegoapi.util.Signature;
 import lombok.Getter;
 
 /**
